+++ spotifywebapipython/__init__.py
@@ -15,6 +15,7 @@ from spotifywebapipython.spotifyapierror import SpotifyApiError
 from spotifywebapipython.spotifyapimessage import SpotifyApiMessage
 from spotifywebapipython.spotifyauthtoken import SpotifyAuthToken
 from spotifywebapipython.spotifyclient import SpotifyClient
+from spotifywebapipython.spotifydiscovery import SpotifyDiscovery
 from spotifywebapipython.spotifywebapiauthenticationerror import SpotifyWebApiAuthenticationError
 from spotifywebapipython.spotifywebapierror import SpotifyWebApiError
 from spotifywebapipython.sautils import GetUnixTimestampMSFromUtcNow
@@ -25,6 +26,7 @@ __all__ = [
     'SpotifyApiMessage',
     'SpotifyAuthToken',
     'SpotifyClient',
+    'SpotifyDiscovery',
     'SpotifyWebApiAuthenticationError',
     'SpotifyWebApiError',
     'GetUnixTimestampMSFromUtcNow'